[BE] 🚚 RENAME: request dto 폴더 생성, 파일 이동
